add swagger, fixed query repository and ngrok setup

+++ src/main/java/com/Merlin/Inventory/Management/System/Product/ProductDto.java
@@ -6,7 +6,7 @@ public record ProductDto(
         String productName,
         String description,
         BigDecimal sellingPrice,
-        Integer minimumQuantity ,
+        Double minimumQuantity ,
         Long supplierId,
         Long categoryId
 ) {
